+++ src/Services/Segments/DTO/SegmentViewDTO.cs
@@ -7,8 +7,8 @@
         public string Key { get; set; }
         public string Avatar { get; set; }
 
-        public float TokensEarned { get; set; }
-        public float TokensSpent { get; set; }
+        public double TokensEarned { get; set; }
+        public double TokensSpent { get; set; }
         public int ChallengesActive { get; set; }
         public int ChallengesCompleted { get; set; }
         public int ShopItemsBought { get; set; }
